Implement new event logging, not very clean yet but it generally works okayish

+++ contracts/gameplay-v1/moves/damage/ControlMove.sol
@@ -49,7 +49,7 @@ contract ControlMove is MoveV1 {
             input.attacker.tokenId,
             input.defender.tokenId,
             address(this),
-            damage,
+            int16(damage),
             elementalMultiplier,
             isCriticalHit
         );
