+++ contracts/gameplay-v1/moves/damage/DamageOverTimeMove.sol
@@ -54,7 +54,7 @@ contract DamageOverTimeMove is MoveV1 {
             input.attacker.tokenId,
             input.defender.tokenId,
             address(this),
-            damage,
+            int16(damage),
             elementalMultiplier,
             isCriticalHit
         );
