+++ contracts/gameplay-v1/moves/damage/PurgeBuffsMove.sol
@@ -48,7 +48,7 @@ contract PurgeBuffsMove is MoveV1 {
             input.attacker.tokenId,
             input.defender.tokenId,
             address(this),
-            damage,
+            int16(damage),
             elementalMultiplier,
             isCriticalHit
         );
